Templates next (#284)
* Add checkout_ui_extension_next and product_subscription_next extensions

* Add checkout_post_purchase_next template

* Some fixes around breaklines

* Make uuids unique

* Add conditional workflow only for the module identifier

* Fix imports

* Remove .tpl from the extension

* Rename to .UsesNext

* Some fixes

* Remove duplicated import

* Generate the right type

* Use jsx when using React in Shopify CLI 3.0

* Fix templates

* Revert changes in rules.go

* Add name field to shopify.extension.toml

* Use explicit node_executable for develop

Also adds an explicit directory for building

* Build extensions from @shopify/app (for now)

* Allow using the expected relative filepaths

* Use node_executable actually present in build project

* Upgrade @shopify/app to 2.0

* Make node_executable unnecessary, run everything through package manager

This is a much stabler interface and will make the results less subejct
to user environment quirks.

* Refactor: Extract shared code to a function

* Fix broken tests

* Symlink shopify-cli-extensions package executable

It's not generated automatically at the repo root node_modules, so we need to
generate explicitly.

* Revert "Refactor: Extract shared code to a function"

This reverts commit b89e9a12051d256b428a5e65e3eae8a54bc8b65b.

* Revert "Make node_executable unnecessary, run everything through package manager"

This reverts commit d68cde994b98490e24fd15d6de1cd572840c3b98.

* Restore deleted test

It was previously deleted in a refactor; as that refactor has been
undone, the test is necessary again.

* More precise implementation
