+++ core/core.go
@@ -148,6 +148,14 @@ type Development struct {
 	Template           string            `json:"-" yaml:"template,omitempty"`
 }
 
+func (e Extension) UsesNext() bool {
+	return strings.HasSuffix(e.Type, "_next")
+}
+
+func (e Extension) NormalizedType() string {
+	return strings.Replace(e.Type, "_next", "", -1)
+}
+
 func (d Development) UsesReact() bool {
 	return strings.Contains(d.Template, "react")
 }
