+++ create/templates/projects/checkout_post_purchase_next/shopify.extension.toml.tpl
@@ -1,0 +1,9 @@
+{{ template "shared/shopify.extension.toml" . }}
+
+# [[metafields]]
+# namespace: my-namespace
+# key: my-key
+
+# [[metafields]]
+# namespace: my-namespace
+# key: my-key-2
