+++ create/templates/projects/checkout_ui_extension_next/locales/en.default.json
@@ -1,0 +1,3 @@
+{
+  "welcome": "Welcome to the {{extensionPoint}} extension!"
+}
