+++ create/templates/projects/checkout_ui_extension_next/locales/fr.json
@@ -1,0 +1,3 @@
+{
+  "welcome": "Bienvenue dans l'extension {{extensionPoint}}!"
+}
