+++ create/templates/projects/checkout_ui_extension_next/shopify.extension.toml.tpl
@@ -1,0 +1,13 @@
+{{- template "shared/shopify.extension.toml" . }}
+
+extension_points = [
+  'Checkout::Dynamic::Render'
+]
+
+# [[metafields]]
+# namespace: my-namespace
+# key: my-key
+
+# [[metafields]]
+# namespace: my-namespace
+# key: my-key-2
