+++ create/templates/projects/checkout_ui_extension_next/src/index.tpl
