+++ create/templates/projects/product_subscription_next/shopify.extension.toml.tpl
@@ -1,0 +1,1 @@
+{{- template "shared/shopify.extension.toml" . -}}
