+++ create/templates/projects/product_subscription_next/src/index.tpl
