+++ create/templates/shared/checkout_post_purchase/javascript.body.js
@@ -1,0 +1,97 @@
+/**
+ * Entry point for the `ShouldRender` Extension Point.
+ *
+ * Returns a value indicating whether or not to render a PostPurchase step, and
+ * optionally allows data to be stored on the client for use in the `Render`
+ * extension point.
+ */
+ extend("Checkout::PostPurchase::ShouldRender", async ({ storage }) => {
+    const initialState = await getRenderData();
+    const render = true;
+  
+    if (render) {
+      // Saves initial state, provided to `Render` via `storage.initialData`
+      await storage.update(initialState);
+    }
+  
+    return {
+      render,
+    };
+  });
+  
+  // Simulate results of network call, etc.
+  async function getRenderData() {
+    return {
+      couldBe: "anything",
+    };
+  }
+  
+  /**
+   * Entry point for the `Render` Extension Point
+   *
+   * Returns markup composed of remote UI components.  The Render extension can
+   * optionally make use of data stored during `ShouldRender` extension point to
+   * expedite time-to-first-meaningful-paint.
+   */
+  extend(
+    "Checkout::PostPurchase::Render",
+    (root, { extensionPoint, storage }) => {
+      const initialState = storage.initialData;
+  
+      root.appendChild(
+        root.createComponent(BlockStack, { spacing: "loose" }, [
+          root.createComponent(
+            CalloutBanner,
+            { title: "Post-purchase extension template" },
+            "Use this template as a starting point to build a great post-purchase extension."
+          ),
+          root.createComponent(
+            Layout,
+            {
+              maxInlineSize: 0.95,
+              media: [
+                { viewportSize: "small", sizes: [1, 30, 1] },
+                { viewportSize: "medium", sizes: [300, 30, 0.5] },
+                { viewportSize: "large", sizes: [400, 30, 0.33] },
+              ],
+            },
+            [
+              root.createComponent(View, {}, [
+                root.createComponent(Image, {
+                  source:
+                    "https://cdn.shopify.com/static/images/examples/img-placeholder-1120x1120.png",
+                }),
+              ]),
+              root.createComponent(View),
+              root.createComponent(BlockStack, { spacing: "xloose" }, [
+                root.createComponent(TextContainer, {}, [
+                  root.createComponent(Heading, {}, "Post-purchase extension"),
+                  root.createComponent(
+                    TextBlock,
+                    {},
+                    "Here you can cross-sell other products, request a product review based on a previous purchase, and much more."
+                  ),
+                ]),
+                root.createComponent(
+                  Button,
+                  {
+                    submit: true,
+                    onPress: () => {
+                      // eslint-disable-next-line
+                      console.log(
+                        `Extension point ${extensionPoint}`,
+                        initialState
+                      );
+                    },
+                  },
+                  "Primary button"
+                ),
+              ]),
+            ]
+          ),
+        ])
+      );
+  
+      root.mount();
+    }
+  );
