+++ create/templates/shared/checkout_post_purchase/react.body.js
@@ -1,0 +1,81 @@
+/**
+ * Entry point for the `ShouldRender` Extension Point.
+ *
+ * Returns a value indicating whether or not to render a PostPurchase step, and
+ * optionally allows data to be stored on the client for use in the `Render`
+ * extension point.
+ */
+ extend("Checkout::PostPurchase::ShouldRender", async ({ storage }) => {
+    const initialState = await getRenderData();
+    const render = true;
+  
+    if (render) {
+      // Saves initial state, provided to `Render` via `storage.initialData`
+      await storage.update(initialState);
+    }
+  
+    return {
+      render,
+    };
+});
+  
+// Simulate results of network call, etc.
+async function getRenderData() {
+    return {
+        couldBe: "anything",
+    };
+}
+
+/**
+ * Entry point for the `Render` Extension Point
+ *
+ * Returns markup composed of remote UI components.  The Render extension can
+ * optionally make use of data stored during `ShouldRender` extension point to
+ * expedite time-to-first-meaningful-paint.
+ */
+render("Checkout::PostPurchase::Render", App);
+
+// Top-level React component
+export function App({ extensionPoint, storage }) {
+    const initialState = storage.initialData;
+
+    return (
+        <BlockStack spacing="loose">
+        <CalloutBanner title="Post-purchase extension template">
+            Use this template as a starting point to build a great post-purchase
+            extension.
+        </CalloutBanner>
+        <Layout
+            maxInlineSize={0.95}
+            media={[
+            { viewportSize: "small", sizes: [1, 30, 1] },
+            { viewportSize: "medium", sizes: [300, 30, 0.5] },
+            { viewportSize: "large", sizes: [400, 30, 0.33] },
+            ]}
+        >
+            <View>
+            <Image source="https://cdn.shopify.com/static/images/examples/img-placeholder-1120x1120.png" />
+            </View>
+            <View />
+            <BlockStack spacing="xloose">
+            <TextContainer>
+                <Heading>Post-purchase extension</Heading>
+                <TextBlock>
+                Here you can cross-sell other products, request a product review
+                based on a previous purchase, and much more.
+                </TextBlock>
+            </TextContainer>
+            <Button
+                submit
+                onPress={() => {
+                // eslint-disable-next-line no-console
+                console.log(`Extension point ${extensionPoint}`, initialState);
+                }}
+            >
+                Primary button
+            </Button>
+            </BlockStack>
+        </Layout>
+        </BlockStack>
+    );
+}
