+++ create/templates/shared/checkout_ui_extension/javascript.js
@@ -1,4 +1,4 @@
-import { extend, Text } from "@shopify/checkout-ui-extensions";
+import { extend, Text } from {{ if .UsesNext }}"@shopify/app/ui-extensions/checkout"{{ else }}"@shopify/checkout-ui-extensions"{{ end }};
 
 extend("Checkout::Dynamic::Render", (root, { extensionPoint, i18n }) => {
   root.appendChild(
