+++ create/templates/shared/checkout_ui_extension/react.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {useExtensionApi, render, Text} from '@shopify/checkout-ui-extensions-react';
+import {useExtensionApi, render, Text} from {{ if .UsesNext }}'@shopify/app/ui-extensions/checkout/react'{{ else }}'@shopify/checkout-ui-extensions-react'{{ end }};
 
 render('Checkout::Dynamic::Render', () => <App />);
