+++ create/templates/shared/product_subscription/react.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { render, extend, Text } from "@shopify/admin-ui-extensions-react";
+import { render, extend, Text } from {{ if .UsesNext }}"@shopify/app/ui-extensions/admin/react"{{ else }}"@shopify/admin-ui-extensions-react"{{ end }};
 
 // Your extension must render all four modes
 extend("Admin::Product::SubscriptionPlan::Add", render(App));
