+++ create/templates/shared/shopify.extension.toml
@@ -1,0 +1,2 @@
+type = "{{ .NormalizedType }}"
+name = "{{ .Title }}"
